# server/docs/swagger.yaml
host: 127.0.0.1:1323
info:
  contact:
    name: API Support
  description: WISMED Wisite-API Server. Updated@ 2022-09-03T11:33:31+10:00
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: WISMED WISITE API
  version: "1.0"
paths:
...
  /api/user/avatar:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK - get avatar src base64
        "404":
          description: Fail - avatar is empty
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: get self avatar src as base64
      tags:
      - User
  /api/user/heartbeats:
    patch:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK - heartbeats successfully
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: frequently call this to indicate that front-end user is active.
      tags:
      - User
  /api/user/profile:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK - profile get successfully
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: get user profile
      tags:
      - User
  /api/user/setprofile:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: real name
        in: formData
        name: name
        type: string
      - description: phone number
        in: formData
        name: phone
        type: string
      - description: address
        in: formData
        name: addr
        type: string
      - description: city
        in: formData
        name: city
        type: string
      - description: country
        in: formData
        name: country
        type: string
      - description: personal id type
        in: formData
        name: pidtype
        type: string
      - description: personal id
        in: formData
        name: pid
        type: string
      - description: gender M/F
        in: formData
        name: gender
        type: string
      - description: date of birth
        in: formData
        name: dob
        type: string
      - description: job position
        in: formData
        name: position
        type: string
      - description: title
        in: formData
        name: title
        type: string
      - description: employer
        in: formData
        name: employer
        type: string
      - description: biography
        in: formData
        name: bio
        type: string
      - description: avatar
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK - profile set successfully
        "400":
          description: Fail - invalid set fields
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: set user profile
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: authorization
    type: apiKey
swagger: "2.0"